Move tpbot specific classes out of support.
