+++ TelepresenceBot/core/src/main/java/com/novoda/tpbot/Event.java
@@ -1,4 +1,4 @@
-package com.novoda.tpbot.support;
+package com.novoda.tpbot;
 
 public enum Event {
