+++ TelepresenceBot/core/src/main/java/com/novoda/tpbot/MalformedServerAddressException.java
@@ -1,4 +1,4 @@
-package com.novoda.tpbot.support;
+package com.novoda.tpbot;
 
 import java.io.IOException;
