+++ TelepresenceBot/core/src/main/java/com/novoda/tpbot/Room.java
@@ -1,0 +1,17 @@
+package com.novoda.tpbot;
+
+public enum Room {
+
+    LONDON("room=London");
+
+    private final String rawQuery;
+
+    Room(String rawQuery) {
+        this.rawQuery = rawQuery;
+    }
+
+    public String rawQuery() {
+        return rawQuery;
+    }
+
+}
