--- TelepresenceBot/core/src/main/java/com/novoda/tpbot/support/ClientType.java
@@ -1,18 +1,0 @@
-package com.novoda.tpbot.support;
-
-public enum ClientType {
-
-    BOT("clientType=bot"),
-    HUMAN("clientType=human");
-
-    private final String rawQuery;
-
-    ClientType(String rawQuery) {
-        this.rawQuery = rawQuery;
-    }
-
-    public String rawQuery() {
-        return rawQuery;
-    }
-
-}
