+++ TelepresenceBot/mobile/src/main/java/com/novoda/tpbot/bot/BotActivity.java
@@ -24,7 +24,7 @@ import com.novoda.tpbot.controls.ControllerListener;
 import com.novoda.tpbot.controls.ControllerView;
 import com.novoda.tpbot.human.ServerDeclarationView;
 import com.novoda.tpbot.support.SelfDestructingMessageView;
-import com.novoda.tpbot.support.ServerDeclarationListener;
+import com.novoda.tpbot.ServerDeclarationListener;
 import com.novoda.tpbot.support.SwitchableView;
 
 import java.util.HashMap;
