+++ TelepresenceBot/mobile/src/main/java/com/novoda/tpbot/human/HumanActivity.java
@@ -11,7 +11,7 @@ import com.novoda.tpbot.controls.CommandRepeater;
 import com.novoda.tpbot.controls.ControllerListener;
 import com.novoda.tpbot.controls.ControllerView;
 import com.novoda.tpbot.support.SelfDestructingMessageView;
-import com.novoda.tpbot.support.ServerDeclarationListener;
+import com.novoda.tpbot.ServerDeclarationListener;
 import com.novoda.tpbot.support.SwitchableView;
 
 public class HumanActivity extends AppCompatActivity implements HumanView {
