+++ TelepresenceBot/mobile/src/main/java/com/novoda/tpbot/human/ServerDeclarationView.java
@@ -8,7 +8,7 @@ import android.widget.LinearLayout;
 
 import com.novoda.notils.caster.Views;
 import com.novoda.tpbot.R;
-import com.novoda.tpbot.support.ServerDeclarationListener;
+import com.novoda.tpbot.ServerDeclarationListener;
 
 public class ServerDeclarationView extends LinearLayout {
